Feat/consume courier assignment event (#23)
* feat: add courier assignment event consumer (not working yet)

* fix: changed type of price to be BigDecimal

* fix: removed config folder

* update files to pass check-style

* update files to pass check-style

* update files to pass check-style

* fix: fix bug for otp & qr confirmation strategies

* update files to pass check-style

---------

+++ src/main/java/com/podzilla/courier/dtos/delivery_tasks/SubmitCourierRatingRequestDto.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class SubmitCourierRatingRequestDto {
@@ -13,5 +15,5 @@ public class SubmitCourierRatingRequestDto {
     @NotNull(message = "Rating is required")
     @Min(value = 1, message = "Rating must be at least 1")
     @Max(value = 5, message = "Rating must be at most 5")
-    private Double rating;
+    private BigDecimal rating;
 }
